+++ src/app/g3w-ol/controls/geocodingcontrol.js
@@ -1,0 +1,614 @@
+const Control = require('./control');
+const { toRawType, XHR } = require('core/utils/utils');
+/**
+ * Classes for all element of dom control
+ * @type {{country: string, hidden: string, city: string, road: string, spin: string, namespace: string, inputText: {container: string, result: string, input: string, reset: string, control: string}, olControl: string, inputResetId: string, inputQueryId: string}}
+ */
+const cssClasses = {
+  namespace: "ol-geocoder",
+  spin: "gcd-pseudo-rotate",
+  hidden: "gcd-hidden",
+  inputQueryId: "gcd-input-query",
+  inputResetId: "gcd-input-reset",
+  country: "gcd-country",
+  city: "gcd-city",
+  road: "gcd-road",
+  olControl: "ol-control",
+  inputText: {
+    container: "gcd-txt-container",
+    control: "gcd-txt-control",
+    input: "gcd-txt-input",
+    reset: "gcd-txt-reset",
+    result: "gcd-txt-result"
+  }
+};
+
+const utils = {
+  isNumeric(str) {
+    return /^\d+$/.test(str);
+  },
+  classRegex(classname) {
+    return new RegExp(("(^|\\s+) " + classname + " (\\s+|$)"));
+  },
+  /**
+   * @param {Element|Array<Element>} element DOM node or array of nodes.
+   * @param {String|Array<String>} classname Class or array of classes.
+   * For example: 'class1 class2' or ['class1', 'class2']
+   * @param {Number|undefined} timeout Timeout to remove a class.
+   */
+  addClass(element, classname, timeout) {
+    if (Array.isArray(element)) {
+      element.forEach(each => this.addClass(each, classname));
+      return;
+    }
+    const array = (Array.isArray(classname))
+      ? classname
+      : classname.split(/\s+/);
+    let i = array.length;
+
+    while (i--) {
+      if (!this.hasClass(element, array[i])) this._addClass(element, array[i], timeout);
+    }
+  },
+  _addClass(el, klass, timeout) {
+    // use native if available
+    if (el.classList) el.classList.add(klass);
+    else el.className = (el.className + ' ' + klass).trim();
+    if (timeout && this.isNumeric(timeout)) setTimeout( () => this._removeClass(el, klass), timeout);
+  },
+  /**
+   * @param {Element|Array<Element>} element DOM node or array of nodes.
+   * @param {String|Array<String>} classname Class or array of classes.
+   * For example: 'class1 class2' or ['class1', 'class2']
+   * @param {Number|undefined} timeout Timeout to add a class.
+   */
+  removeClass(element, classname, timeout) {
+    if (Array.isArray(element)) {
+      element.forEach(each => this.removeClass(each, classname, timeout));
+      return;
+    }
+    const array = (Array.isArray(classname))
+      ? classname
+      : classname.split(/\s+/);
+    let i = array.length;
+
+    while (i--) {
+      if (this.hasClass(element, array[i])) this._removeClass(element, array[i], timeout);
+    }
+  },
+  _removeClass(el, klass, timeout) {
+    if (el.classList) el.classList.remove(klass);
+    else el.className = (el.className.replace(this.classRegex(klass), ' ')).trim();
+    if (timeout && this.isNumeric(timeout)) setTimeout( () => this._addClass(el, klass), timeout);
+  },
+  /**
+   * @param {Element} element DOM node.
+   * @param {String} classname Classname.
+   * @return {Boolean}
+   */
+  hasClass(element, c) {
+    // use native if available
+    return element.classList ? element.classList.contains(c) : this.classRegex(c).test(element.className);
+  },
+  /**
+   * @param {Element|Array<Element>} element DOM node or array of nodes.
+   * @param {String} classname Classe.
+   */
+  toggleClass(element, classname) {
+    if (Array.isArray(element)) {
+      element.forEach( each => this.toggleClass(each, classname));
+      return;
+    }
+    // use native if available
+    if (element.classList) element.classList.toggle(classname);
+    else {
+      if (this.hasClass(element, classname)) this._removeClass(element, classname);
+      else this._addClass(element, classname);
+    }
+  },
+  /**
+   * Abstraction to querySelectorAll for increased
+   * performance and greater usability
+   * @param {String} selector
+   * @param {Element} context (optional)
+   * @param {Boolean} find_all (optional)
+   * @return (find_all) {Element} : {Array}
+   */
+  find(selector, context, find_all) {
+    if(context === void 0) context = window.document;
+    let simpleRe = /^(#?[\w-]+|\.[\w-.]+)$/,
+      periodRe = /\./g,
+      slice = Array.prototype.slice,
+      matches = [];
+    // Redirect call to the more performant function
+    // if it's a simple selector and return an array
+    // for easier usage
+    if (simpleRe.test(selector)) {
+      switch (selector[0]) {
+        case '#':
+          matches = [this.$(selector.substr(1))];
+          break;
+        case '.':
+          matches = slice.call(context.getElementsByClassName(
+            selector.substr(1).replace(periodRe, ' ')));
+          break;
+        default:
+          matches = slice.call(context.getElementsByTagName(selector));
+      }
+    } else {
+      // If not a simple selector, query the DOM as usual
+      // and return an array for easier usage
+      matches = slice.call(context.querySelectorAll(selector));
+    }
+    return find_all ? matches : matches[0];
+  },
+  $(id) {
+    id = (id[0] === '#') ? id.substr(1, id.length) : id;
+    return document.getElementById(id);
+  },
+  isElement(obj) {
+    // DOM, Level2
+    if ('HTMLElement' in window) return (!!obj && obj instanceof HTMLElement);
+    // Older browsers
+    return (!!obj && typeof obj === 'object' && obj.nodeType === 1 &&
+      !!obj.nodeName);
+  },
+  getAllChildren(node, tag) {
+    return [].slice.call(node.getElementsByTagName(tag));
+  },
+  isEmpty(str) {
+    return (!str || 0 === str.length);
+  },
+  emptyArray(array) {
+    while (array.length) {
+      array.pop();
+    }
+  },
+  anyMatchInArray(source, target) {
+    return source.some(each => target.indexOf(each) >= 0);
+  },
+  everyMatchInArrayy(arr1, arr2) {
+    return arr2.every(each => arr1.indexOf(each) >= 0);
+  },
+  anyItemHasValue(obj, has) {
+    if (has === undefined) has = false;
+    const keys = Object.keys(obj);
+    keys.forEach(key=>  {
+      if (!this.isEmpty(obj[key])) {
+        has = true;
+      }
+    });
+    return has;
+  },
+  removeAllChildren(node) {
+    while (node.firstChild) {
+      node.removeChild(node.firstChild);
+    }
+  },
+  removeAll(collection) {
+    let node;
+    while ((node = collection[0])) {
+      node.parentNode.removeChild(node);
+    }
+  },
+  getChildren(node, tag) {
+    return [].filter.call(
+      node.childNodes, function (el) {
+        return tag
+          ? el.nodeType === 1 && el.tagName.toLowerCase() === tag
+          : el.nodeType === 1;
+      }
+    );
+  },
+  template(html, row) {
+    return html.replace(/\{ *([\w_-]+) *\}/g, (htm, key) => {
+      const value = (row[key] === undefined) ? '' : row[key];
+      return this.htmlEscape(value);
+    });
+  },
+  htmlEscape(str) {
+    return String(str)
+      .replace(/&/g, '&amp;')
+      .replace(/</g, '&lt;')
+      .replace(/>/g, '&gt;')
+      .replace(/"/g, '&quot;')
+      .replace(/'/g, '&#039;');
+  },
+  createElement(node, html) {
+    let elem;
+    if (Array.isArray(node)) {
+      elem = document.createElement(node[0]);
+      if (node[1].id) elem.id = node[1].id;
+      if (node[1].classname) elem.className = node[1].classname;
+      if (node[1].attr) {
+        const attr = node[1].attr;
+        if (Array.isArray(attr)) {
+          let i = -1;
+          while (++i < attr.length) {
+            elem.setAttribute(attr[i].name, attr[i].value);
+          }
+        } else elem.setAttribute(attr.name, attr.value);
+      }
+    } else elem = document.createElement(node);
+    elem.innerHTML = html;
+    const frag = document.createDocumentFragment();
+
+    while (elem.childNodes[0]) {
+      frag.appendChild(elem.childNodes[0]);
+    }
+    elem.appendChild(frag);
+    return elem;
+  }
+};
+
+/**
+ * Providers
+ */
+class Nominatim {
+  constructor(options={}) {
+    this.settings = {
+      url: 'https://nominatim.openstreetmap.org/search/',
+      params: {
+        q: '',
+        format: 'json',
+        addressdetails: 1,
+        limit: 10
+      },
+      viewbox: ol.proj.transformExtent(options.viewbox, options.mapCrs, 'EPSG:4326').join(',')
+    };
+  }
+  getParameters(options) {
+    const {url, viewbox, params:{limit}} = this.settings;
+    return {
+      url,
+      params: {
+        q: options.query,
+        format: 'json',
+        addressdetails: 1,
+        limit: options.limit || limit,
+        viewbox,
+        bounded: 1
+      }
+    };
+  };
+  handleResponse(results) {
+    return results.map(result => ({
+        lon: result.lon,
+        lat: result.lat,
+        address: {
+          name: result.address.neighbourhood || '',
+          road: result.address.road || '',
+          postcode: result.address.postcode,
+          city: result.address.city || result.address.town,
+          state: result.address.state,
+          country: result.address.country
+        },
+        original: {
+          formatted: result.display_name,
+          details: result.address
+        }
+      })
+    )
+  }
+}
+
+/**
+ * Geocoding class
+ * @param options
+ * @constructor
+ */
+function GeocodingControl(options={}) {
+  this.options = {
+    provider: 'osm',
+    placeholder: options.placeholder || 'Città, indirizzo ... ',
+    noresults: options.noresults || 'Nessun risultato ',
+    notresponseserver: options.notresponseserver || 'Il server non risponde',
+    lang: 'it-IT',
+    limit: 5,
+    keepOpen: true,
+    preventDefault: false,
+    autoComplete: false,
+    autoCompleteMinLength: 4,
+    debug: false,
+    viewbox: options.bbox,
+    bounded: 1,
+    classMobile: options.isMobile ? 'nominatim-mobile' : '',
+    mapCrs: options.mapCrs,
+    fontIcon: options.fontIcon || "fa fa-search fas fa-search"
+  };
+  /**
+   * set currentprovider
+   * @type {string|TableLayer.providers.data}
+   */
+  this.currentProvider = options.provider || 'nominatim';
+  const {placeholder, fontIcon, mapCrs, viewbox} = this.options;
+  /**
+   * Providers
+   * @type {{nominatim: Nominatim, google: null}}
+   */
+  this.providers = {
+    nominatim: new Nominatim({
+      mapCrs,
+      viewbox
+    }),
+    google: null // TO DO
+  };
+  const containerClass = `${cssClasses.namespace} ${cssClasses.inputText.container} ${this.options.classMobile}`;
+  const self = this;
+  const nominatimVueContainer = Vue.extend({
+    functional: true,
+    render(h){
+      return h('div', {class: {[containerClass]: true}}, [
+        h('div', {
+          class: {
+            [cssClasses.inputText.control]: true,
+          }
+        }, [
+          h('input', {
+            attrs: {
+              type: 'text',
+              id: cssClasses.inputQueryId,
+              autocomplete: 'off'
+            },
+            class:{
+              [cssClasses.inputText.input]: true
+            },
+            directives:[
+              {
+                name: 't-placeholder',
+                value: placeholder
+              }
+            ]
+          }),
+          h('button', {
+            attrs: {
+              type: 'button',
+              id: 'search_nominatim'
+            },
+            class:{
+              btn: true
+            },
+            on: {
+              click() {
+                const value = $(`input.${cssClasses.inputText.input}`).val();
+                self.query(value);
+              }
+            }
+          }, [h('i', {
+            attrs: {
+              'aria-hidden': true
+            },
+            style: {
+              color:'#ffffff'
+            },
+            class: {
+              [fontIcon]: true
+            }
+          })]),
+          h('button', {
+            attrs: {
+              type: 'button',
+              id:  cssClasses.inputResetId
+            },
+            class: {
+              [`${cssClasses.inputText.reset}  ${cssClasses.hidden}`]: true
+            }
+          }),
+        ]),
+        h('ul', {
+          class: {
+            [cssClasses.inputText.result]: true
+          }
+        })])
+    }
+  });
+  this.container = new nominatimVueContainer().$mount().$el;
+  this.lastQuery = '';
+  this.registeredListeners = {
+    mapClick: false
+  };
+  /**
+   * Methods
+   */
+  this.createControl = function() {
+    let timeout, lastQuery;
+    this.control = utils.find('.' + cssClasses.inputText.control, this.container);
+    this.input = utils.find('.' + cssClasses.inputText.input, this.container);
+    this.reset = utils.find('.' + cssClasses.inputText.reset, this.container);
+    this.result = utils.find('.' + cssClasses.inputText.result, this.container);
+    const query =  evt => {
+      const value = evt.target.value.trim();
+      const hit = evt.key ? evt.key === 'Enter' : evt.which ? evt.which === 13 : evt.keyCode ? evt.keyCode === 13 : false;
+      if (hit) {
+        evt.preventDefault();
+        this.query(value);
+      }
+    };
+    const reset = () => {
+      this.input.focus();
+      this.input.value = '';
+      this.lastQuery = '';
+      utils.addClass(this.reset, cssClasses.hidden);
+      this.clearResults();
+    };
+    const handleValue =  evt => {
+      const value = evt.target.value.trim();
+      value.length ? utils.removeClass(this.reset, cssClasses.hidden) : utils.addClass(this.reset, cssClasses.hidden);
+      if (this.options.autoComplete && value !== lastQuery) {
+        lastQuery = value;
+        timeout && clearTimeout(timeout);
+        timeout = setTimeout( () =>  (value.length >= this.options.autoCompleteMinLength) && this.query(value), 200);
+      }
+    };
+    this.input.addEventListener('keyup', query, false);
+    this.input.addEventListener('input', handleValue, false);
+    this.reset.addEventListener('click', reset, false);
+  };
+  this.query = function(q) {
+    return new Promise((resolve, reject) => {
+      const isNumber = value => toRawType(value) === 'Number' && !Number.isNaN(value);
+      let lonlat = null;
+      if (q && q.split(',').length === 2) {
+        lonlat = q.split(',');
+        lonlat = isNumber(1*lonlat[0]) && isNumber(1*lonlat[1]) ? lonlat.map(coordinate => 1*coordinate) : null;
+      }
+      if (lonlat) this.dispatchEvent({
+        type: 'lonlat',
+        lonlat
+      });
+      else {
+        if (this.lastQuery === q && this.result.firstChild) { return; }
+        const provider =  this.providers[this.currentProvider]
+        const {url, params} = provider.getParameters({
+          query: q,
+          key: this.options.key,
+          lang: this.options.lang,
+          countrycodes: this.options.countrycodes,
+          limit: this.options.limit
+        });
+        this.lastQuery = q;
+        this.clearResults();
+        utils.addClass(this.reset, cssClasses.spin);
+        XHR.get({
+          url,
+          params
+        })
+          .then(res => {
+            const extent = params.viewbox.split(',').map(coordinate => 1*coordinate);
+            res = res.filter(place => ol.extent.containsXY(extent, place.lon, place.lat));
+            utils.removeClass(this.reset, cssClasses.spin);
+            const res_= res.length ? provider.handleResponse(res) : undefined;
+            this.createList(res_);
+            res_ && this.listenMapClick();
+            resolve(res_ ? res_ : []);
+          })
+          .catch(error => {
+            utils.removeClass(this.reset, cssClasses.spin);
+            const li = utils.createElement(
+              'li', `<h5>  ${this.options.notresponseserver}</h5>`);
+            this.result.appendChild(li);
+            reject(error)
+          })
+      }
+    })
+  };
+  this.createList = function(response) {
+    const ul = this.result;
+    if (response) {
+      response.forEach(row => {
+        const addressHtml = this.addressTemplate(row.address),
+          html = ['<a href="#">', addressHtml, '</a>'].join(''),
+          li = utils.createElement('li', html);
+        li.addEventListener('click', evt => {
+          evt.preventDefault();
+          this.chosen(row, addressHtml, row.address, row.original);
+        }, false);
+        ul.appendChild(li);
+      });
+    } else {
+      const {noresults} = this.options;
+      const elementVue = Vue.extend({
+        functional: true,
+        render(h){
+          return h('li', {
+            class: {
+              'nominatim-noresult': true
+            },
+            directives:[{name: 't', value: noresults}]
+          })
+        }
+      });
+      const li = new elementVue().$mount().$el;
+      ul.appendChild(li);
+    }
+  };
+  this.chosen = function(place, addressHtml, addressObj, addressOriginal) {
+    const map = this.getMap();
+    const coord_ = [parseFloat(place.lon), parseFloat(place.lat)];
+    const projection = map.getView().getProjection();
+    const coord = ol.proj.transform(coord_, 'EPSG:4326', projection);
+    const address = {
+      formatted: addressHtml,
+      details: addressObj,
+      original: addressOriginal
+    };
+    this.options.keepOpen === false && this.clearResults(true);
+    this.dispatchEvent({
+      type: 'address',
+      address: address,
+      coordinate: coord
+    });
+  };
+  this.addressTemplate = function(address) {
+    const html = [];
+    if (address.name) html.push(['<div class="', cssClasses.road, '">{name}</div>'].join(''));
+    if (address.road || address.building || address.house_number) {
+      html.push([
+        '<div class="', cssClasses.road,
+        '">{building} {road} {house_number}</div>'
+      ].join(''));
+    }
+    if (address.city || address.town || address.village) {
+      html.push([
+        '<div class="', cssClasses.city,
+        '">{postcode} {city} {town} {village}</div>'
+      ].join(''));
+    }
+    if (address.state || address.country) {
+      html.push([
+        '<div class="', cssClasses.country, '">{state} {country}</div>'
+      ].join(''));
+    }
+    return utils.template(html.join('<br>'), address);
+  };
+  this.expand = function() {
+    utils.removeClass(this.input, cssClasses.spin);
+    utils.addClass(this.control, cssClasses.glass.expanded);
+    setTimeout( () =>  this.input.focus(), 100);
+    this.listenMapClick();
+  };
+  this.collapse = function() {
+    this.input.value = '';
+    this.input.blur();
+    utils.addClass(this.reset, cssClasses.hidden);
+    this.clearResults();
+  };
+  this.listenMapClick = function() {
+    // already registered
+    if (this.registeredListeners.mapClick) return;
+    const mapElement = this.getMap().getTargetElement();
+    this.registeredListeners.mapClick = true;
+    //one-time fire click
+    mapElement.addEventListener('click', evt => {
+      this.clearResults(true);
+      mapElement.removeEventListener(evt.type, this, false);
+      this.registeredListeners.mapClick = false;
+    }, false);
+  };
+  this.clearResults = function() {
+    utils.removeAllChildren(this.result);
+  };
+  this.getSource = function() {
+    return this.layer.getSource();
+  };
+  this.addLayer = function() {
+    const map = this.getMap();
+    const layer = map.getLayers().find(layer =>  layer === this.layer);
+    !layer && map.addLayer(this.layer);
+  };
+  /**
+   * End methods
+   */
+
+  this.createControl();
+
+  Control.call(this, {
+    element: this.container,
+    name: "nominatim",
+    offline: false
+  });
+}
+
+ol.inherits(GeocodingControl, Control);
+
+module.exports = GeocodingControl;
