+++ src/app/gui/map/control/factory.js
@@ -9,7 +9,7 @@ const AddLayersControl = require('g3w-ol/controls/addlayers');
 const LengthControl = require('g3w-ol/controls/lengthcontrol');
 const AreaControl = require('g3w-ol/controls/areacontrol');
 const OLControl = require('g3w-ol/controls/olcontrol');
-const NominatimControl = require('g3w-ol/controls/nominatimcontrol');
+const NominatimControl = require('g3w-ol/controls/geocodingcontrol');
 const MousePositionControl = require('g3w-ol/controls/mousepositioncontrol');
 const ScaleControl = require('g3w-ol/controls/scalecontrol');
 const OnClikControl = require('g3w-ol/controls/onclickcontrol');
@@ -18,10 +18,8 @@ const geoScreenshotControl = require('g3w-ol/controls/geoscreenshotcontrol');
 
 const ControlsFactory = {
   create(options={}) {
-    let control;
     const ControlClass = ControlsFactory.CONTROLS[options.type];
-    if (ControlClass) control = new ControlClass(options);
-    return control;
+    if (ControlClass) return new ControlClass(options);
   }
 };
